client-hints/accept-ch-change.tentative.https.html

<html>
<title>Accept-CH test</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<!--
Apart from this webpage, the test opens two more html web page. One test is run
in this web page, and two in the other web pages.
-->

<script>

// This test fetches resources/accept_ch.html. The response headers to
// that webpage contains only the Accept-CH header. These preferences should be
// stored so that the next request to the same origin is sent with the
// requested client hint headers.

// Next, to verify that the origin preferences were persisted by the user
// agent, this test fetches resources/expect_client_hints_headers.html in a new
// window. Fetching of resources/expect_client_hints_headers.html verifies that
// the user agent does send the client hints in the request headers.

// Test is marked as tentative until https://github.com/whatwg/fetch/issues/726
// is resolved.

// First, verify the initial state to make sure that the browser does not have
// client hints preferences cached from a previous run of the test.
promise_test(t => {
  return fetch("echo_client_hints_received.py").then(r => {
    assert_equals(r.status, 200)
    // Verify that the browser did not include client hints in the request
    // headers when fetching echo_client_hints_received.py.
    assert_false(r.headers.has("device-memory-received"), "device-memory-received");
    assert_false(r.headers.has("viewport-width-received"), "viewport-width-received");
  });
}, "Precondition: Test that the browser does not have client hints preferences cached");

function acceptChLoaded() {
  // Open a new window. Verify that the user agent does attach the client hints.
  var verify_win = window.open("resources/expect_client_hints_headers.html");
  assert_not_equals(verify_win, null, "Popup windows not allowed?");
  verify_win.addEventListener('load', changeCh, false);
}

function changeCh() {
  // Use new window to overwrite client hint preferences, then verify new
  // settings have been saved
  var change_ch_win = window.open("resources/accept_ch_different.html");
  assert_not_equals(change_ch_win, null, "Popup windows not allowed?");
  change_ch_win.addEventListener('load', verifyChangeCh, false);
}

function verifyChangeCh() {
  // Use new window to overwrite client hint preferences, then verify new
  // settings have been saved
  var verify_change_win = window.open("resources/expect_client_hints_headers_different.html");
  assert_not_equals(verify_change_win, null, "Popup windows not allowed?");
  verify_change_win.addEventListener('load', clearCh, false);

}

function clearCh() {
  // Use new window to overwrite client hint preferences, then verify new
  // settings have been saved
  var clear_ch_win = window.open("resources/accept_ch_empty.html");
  assert_not_equals(clear_ch_win, null, "Popup windows not allowed?");
  clear_ch_win.addEventListener('load', verifyClearCh, false);
}

function verifyClearCh() {
  // Use new window to overwrite client hint preferences, then verify new
  // settings have been saved
  var verify_clear_win = window.open("resources/do_not_expect_client_hints.html");
  assert_not_equals(verify_change_win, null, "Popup windows not allowed?");
}


// Fetching this webpage should cause user-agent to persist client hint
// preferences for the origin.
var win = window.open("resources/accept_ch.html");
assert_not_equals(win, null, "Popup windows not allowed?");
win.addEventListener('load', acceptChLoaded, false);

</script>
</body>
</html>
